third_party/blink/web_tests/platform/generic/fast/dom/shadow/shadow-element-expected.txt

This test ensures that the shadow element is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadow.nodeName is 'SHADOW'
PASS 0 <= shadow.constructor.toString().indexOf('HTMLShadowElement') is true
PASS window.HTMLShadowElement is shadow.constructor
PASS successfullyParsed is true

TEST COMPLETE
